+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/http/marshaller/JsonSupport.scala
@@ -2,31 +2,26 @@ package io.iohk.atala.agent.server.http.marshaller
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import io.iohk.atala.agent.openapi.model.*
-import spray.json.{DefaultJsonProtocol, RootJsonFormat, jsonReader}
-import spray.json.JsonFormat
+import spray.json.{
+  DefaultJsonProtocol,
+  DeserializationException,
+  JsString,
+  JsValue,
+  JsonFormat,
+  RootJsonFormat,
+  jsonReader,
+  jsonWriter
+}
+
 import java.util.UUID
-import spray.json.JsString
-import spray.json.JsValue
-import spray.json.DeserializationException
 import java.time.OffsetDateTime
 
 trait JsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
 
   // Castor
-  given RootJsonFormat[AuthenticationChallengeSubmissionRequest] = jsonFormat3(
-    AuthenticationChallengeSubmissionRequest.apply
-  )
-  given RootJsonFormat[AuthenticationChallengeSubmissionResponse] = jsonFormat2(
-    AuthenticationChallengeSubmissionResponse.apply
-  )
-  given RootJsonFormat[CreateAuthenticationChallengeRequest] = jsonFormat3(CreateAuthenticationChallengeRequest.apply)
-  given RootJsonFormat[CreateAuthenticationChallengeResponse] = jsonFormat2(CreateAuthenticationChallengeResponse.apply)
   given RootJsonFormat[CreateManagedDidRequest] = jsonFormat1(CreateManagedDidRequest.apply)
   given RootJsonFormat[CreateManagedDidRequestDocumentTemplate] = jsonFormat2(
     CreateManagedDidRequestDocumentTemplate.apply
-  )
-  given RootJsonFormat[CreateManagedDidRequestDocumentTemplatePublicKeysInner] = jsonFormat2(
-    CreateManagedDidRequestDocumentTemplatePublicKeysInner.apply
   )
   given RootJsonFormat[CreateManagedDIDResponse] = jsonFormat1(CreateManagedDIDResponse.apply)
   given RootJsonFormat[DID] = jsonFormat8(DID.apply)
@@ -36,8 +31,20 @@ trait JsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
   given RootJsonFormat[DIDResponse] = jsonFormat2(DIDResponse.apply)
   given RootJsonFormat[ErrorResponse] = jsonFormat5(ErrorResponse.apply)
   given RootJsonFormat[ListManagedDIDResponseInner] = jsonFormat3(ListManagedDIDResponseInner.apply)
+  given RootJsonFormat[ManagedDIDKeyTemplate] = jsonFormat2(ManagedDIDKeyTemplate.apply)
   given RootJsonFormat[PublicKeyJwk] = jsonFormat5(PublicKeyJwk.apply)
   given RootJsonFormat[Service] = jsonFormat3(Service.apply)
+  given RootJsonFormat[UpdateManagedDIDRequest] = jsonFormat1(UpdateManagedDIDRequest.apply)
+  given RootJsonFormat[UpdateManagedDIDRequestActionsInner] = jsonFormat6(UpdateManagedDIDRequestActionsInner.apply)
+  given RootJsonFormat[UpdateManagedDIDRequestActionsInnerRemoveKey] = jsonFormat1(
+    UpdateManagedDIDRequestActionsInnerRemoveKey.apply
+  )
+  given RootJsonFormat[UpdateManagedDIDRequestActionsInnerRemoveService] = jsonFormat1(
+    UpdateManagedDIDRequestActionsInnerRemoveService.apply
+  )
+  given RootJsonFormat[UpdateManagedDIDRequestActionsInnerUpdateService] = jsonFormat3(
+    UpdateManagedDIDRequestActionsInnerUpdateService.apply
+  )
   given RootJsonFormat[VerificationMethod] = jsonFormat4(VerificationMethod.apply)
 
   // Issue Credential Protocol